orchestrator/requirements.txt
blinker==1.6.2
certifi==2023.5.7
charset-normalizer==3.2.0
click==8.1.4
Flask==2.3.2
idna==3.4
interchange==2021.0.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
monotonic==1.6
packaging==23.1
pansi==2020.7.3
py2neo==2021.2.3
Pygments==2.15.1
PyJWT==2.7.0
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0
requests==2.31.0
six==1.16.0
urllib3==2.0.3
Werkzeug==2.3.6
